refactor(state_cache): simplify `StateCache`s by reusing `EstimatedBTreeMap` (#17895)
