+++ src/common/estimate_size/src/collections/btreemap.rs
@@ -18,6 +18,7 @@ use std::ops::{Bound, RangeInclusive};
 
 use crate::{EstimateSize, KvSize};
 
+#[derive(Clone)]
 pub struct EstimatedBTreeMap<K, V> {
     inner: BTreeMap<K, V>,
     heap_size: KvSize,
@@ -42,6 +43,51 @@ impl<K, V> EstimatedBTreeMap<K, V> {
     pub fn is_empty(&self) -> bool {
         self.inner.is_empty()
     }
+
+    pub fn iter(&self) -> impl Iterator<Item = (&K, &V)> {
+        self.inner.iter()
+    }
+
+    pub fn range<R>(&self, range: R) -> std::collections::btree_map::Range<'_, K, V>
+    where
+        K: Ord,
+        R: std::ops::RangeBounds<K>,
+    {
+        self.inner.range(range)
+    }
+
+    pub fn values(&self) -> impl Iterator<Item = &V> {
+        self.inner.values()
+    }
+}
+
+impl<K, V> EstimatedBTreeMap<K, V>
+where
+    K: Ord,
+{
+    pub fn first_key_value(&self) -> Option<(&K, &V)> {
+        self.inner.first_key_value()
+    }
+
+    pub fn first_key(&self) -> Option<&K> {
+        self.first_key_value().map(|(k, _)| k)
+    }
+
+    pub fn first_value(&self) -> Option<&V> {
+        self.first_key_value().map(|(_, v)| v)
+    }
+
+    pub fn last_key_value(&self) -> Option<(&K, &V)> {
+        self.inner.last_key_value()
+    }
+
+    pub fn last_key(&self) -> Option<&K> {
+        self.last_key_value().map(|(k, _)| k)
+    }
+
+    pub fn last_value(&self) -> Option<&V> {
+        self.last_key_value().map(|(_, v)| v)
+    }
 }
 
 impl<K, V> Default for EstimatedBTreeMap<K, V> {
@@ -55,27 +101,23 @@ where
     K: EstimateSize + Ord,
     V: EstimateSize,
 {
-    pub fn first_key_value(&self) -> Option<(&K, &V)> {
-        self.inner.first_key_value()
-    }
-
-    pub fn last_key_value(&self) -> Option<(&K, &V)> {
-        self.inner.last_key_value()
-    }
-
-    pub fn insert(&mut self, key: K, value: V) {
+    pub fn insert(&mut self, key: K, value: V) -> Option<V> {
         let key_size = self.heap_size.add_val(&key);
         self.heap_size.add_val(&value);
-        if let Some(old_value) = self.inner.insert(key, value) {
+        let old_value = self.inner.insert(key, value);
+        if let Some(old_value) = &old_value {
             self.heap_size.sub_size(key_size);
-            self.heap_size.sub_val(&old_value);
+            self.heap_size.sub_val(old_value);
         }
+        old_value
     }
 
-    pub fn remove(&mut self, key: &K) {
-        if let Some(value) = self.inner.remove(key) {
-            self.heap_size.sub(key, &value);
+    pub fn remove(&mut self, key: &K) -> Option<V> {
+        let old_value = self.inner.remove(key);
+        if let Some(old_value) = &old_value {
+            self.heap_size.sub(key, old_value);
         }
+        old_value
     }
 
     pub fn clear(&mut self) {
@@ -100,17 +142,6 @@ where
             inner,
             heap_size: &mut self.heap_size,
         })
-    }
-
-    pub fn iter(&self) -> impl Iterator<Item = (&K, &V)> {
-        self.inner.iter()
-    }
-
-    pub fn range<R>(&self, range: R) -> std::collections::btree_map::Range<'_, K, V>
-    where
-        R: std::ops::RangeBounds<K>,
-    {
-        self.inner.range(range)
     }
 
     /// Retain the given range of entries in the map, removing others.
