+++ src/stream/src/common/mod.rs
@@ -14,10 +14,10 @@
 
 pub use column_mapping::*;
 
-pub mod cache;
 mod column_mapping;
 pub mod compact_chunk;
 pub mod log_store_impl;
 pub mod metrics;
 pub mod rate_limit;
+pub mod state_cache;
 pub mod table;
