+++ src/stream/src/common/table/state_table_cache.rs
@@ -17,7 +17,7 @@ use risingwave_common::row::{OwnedRow, Row, RowExt};
 use risingwave_common::types::{DefaultOrdered, ScalarRefImpl};
 use risingwave_common_estimate_size::EstimateSize;
 
-use crate::common::cache::{StateCache, TopNStateCache};
+use crate::common::state_cache::{StateCache, TopNStateCache};
 
 /// The watermark cache key is just an `OwnedRow` wrapped in `DefaultOrdered`.
 /// This is because we want to use the `DefaultOrdered` implementation of `Ord`.
@@ -125,7 +125,7 @@ impl StateTableWatermarkCache {
     }
 
     pub fn capacity(&self) -> usize {
-        self.inner.capacity_inner()
+        self.inner.capacity()
     }
 
     pub fn len(&self) -> usize {
@@ -186,7 +186,7 @@ mod tests {
     use risingwave_common::types::{Scalar, Timestamptz};
 
     use super::*;
-    use crate::common::cache::StateCacheFiller;
+    use crate::common::state_cache::StateCacheFiller;
 
     /// With capacity 3, test the following sequence of inserts:
     /// Insert
